fontscan: move go1.16+ path traversal into separate files
This commit migrates the more efficient filesystem walking functionality
into a separate file guarded by a build constraint so that the pacakge can
still compile on go 1.14.
