+++ fontscan/scan.go
@@ -4,7 +4,6 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
-	"io/fs"
 	"log"
 	"os"
 	"os/user"
@@ -157,45 +156,7 @@ func ignoreFontFile(name string) bool {
 }
 
 type fontFileHandler interface {
-	consume(path string, info fs.FileInfo) error
-}
-
-// recursively walk through the given directory, scanning font files and calling dst.consume
-// for each valid file found.
-func scanDirectory(dir string, visited map[string]bool, dst fontFileHandler) error {
-	walkFn := func(path string, d fs.DirEntry, err error) error {
-		if err != nil {
-			return fmt.Errorf("error walking font directories: %s", err)
-		}
-
-		if d.IsDir() { // keep going
-			return nil
-		}
-
-		if visited[path] {
-			return nil // skip the path
-		}
-		visited[path] = true
-
-		// load the information, following potential symoblic links
-		info, err := os.Stat(path)
-		if err != nil {
-			return err
-		}
-
-		// always ignore files which should never be font files
-		if ignoreFontFile(info.Name()) {
-			return nil
-		}
-
-		err = dst.consume(path, info)
-
-		return err
-	}
-
-	err := filepath.WalkDir(dir, walkFn)
-
-	return err
+	consume(path string, info os.FileInfo) error
 }
 
 // --------------------- footprint mode -----------------------
@@ -269,7 +230,7 @@ func newFootprintAccumulator(currentIndex systemFontsIndex) footprintScanner {
 	return out
 }
 
-func (fa *footprintScanner) consume(path string, info fs.FileInfo) error {
+func (fa *footprintScanner) consume(path string, info os.FileInfo) error {
 	modTime := newTimeStamp(info)
 
 	// try to avoid scanning the file
@@ -366,7 +327,7 @@ func isFontFile(fileName string) bool {
 	}
 }
 
-func (fns *fileNameScanner) consume(path string, _ fs.FileInfo) error {
+func (fns *fileNameScanner) consume(path string, _ os.FileInfo) error {
 	if isFontFile(path) {
 		*fns = append(*fns, path)
 	}
